basic reinhard tone mapping (with separated color axes) working, but average luminance is not

+++ src/events.rs
@@ -5,7 +5,7 @@ use sdl2::mouse::RelativeMouseState;
 pub fn handle_keyboard(
     game_state: &mut GameState,
     scancodes: Vec<(sdl2::keyboard::Scancode, bool)>,
-    dt: u128,
+    dt: f32,
 ) {
     let camera_movement = scancodes
         .iter()
@@ -28,13 +28,13 @@ pub fn handle_keyboard(
     game_state.move_camera_by_vector(camera_movement, dt);
 }
 
-pub fn handle_mouse(game_state: &mut GameState, mouse_state: &RelativeMouseState, dt: u128) {
+pub fn handle_mouse(game_state: &mut GameState, mouse_state: &RelativeMouseState, dt: f32) {
     let yo = mouse_state.y() as f32;
     let xo = mouse_state.x() as f32;
     game_state.rotate_camera(glam::vec3(yo, -xo, 0.0), dt);
 }
 
-pub fn handle_event(game_state: &mut GameState, event: GameStateEvent, dt: u128) {
+pub fn handle_event(game_state: &mut GameState, event: GameStateEvent, dt: f32) {
     match event {
         GameStateEvent::FrameEvent(scancodes, mouse_state) => {
             handle_keyboard(game_state, scancodes, dt);
